+++ src/sub_commands/mod.rs
@@ -3,6 +3,7 @@ pub mod broadcast_events;
 pub mod convert_key;
 pub mod create_badge;
 pub mod create_public_channel;
+pub mod custom_event;
 pub mod delete_event;
 pub mod dm;
 pub mod generate_keypair;