+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add option to print keys as hex instead of bech32 by [@thesimplekid](https://github.com/thesimplekid)
 - Add LUD-16 to update metadata command by [@w3irdrobot](https://github.com/w3irdrobot)
 - Add support for creating NIP-57 events (Zaps) by [@0xtrr](https://github.com/0xtrr)
+- Add support for custom events with arbitrary kind, content and tags by [@0xtrr](https://github.com/0xtrr)
 
 ### Changed
 - Upgrade nostr-sdk from 0.18 to 0.20 by [@thesimplekid](https://github.com/thesimplekid)